Former-commit-id: ebb2ac96e8bc646c5bf2843b0722993a4583eebc
Former-commit-id: b0d45da51b4372598357ea5565386fa8314ef5be

+++ concrete/src/Tree/Node/Node.php
@@ -6,6 +6,7 @@ use Concrete\Core\Permission\Access\Access;
 use Concrete\Core\Permission\Access\Entity\GroupCombinationEntity;
 use Concrete\Core\Permission\Access\Entity\GroupEntity;
 use Concrete\Core\Permission\Access\Entity\UserEntity;
+use Concrete\Core\Permission\AssignableObjectInterface;
 use Concrete\Core\Permission\AssignableObjectTrait;
 use Concrete\Core\Permission\Key\Key;
 use Concrete\Core\Permission\Key\TreeNodeKey;
@@ -21,7 +22,7 @@ use stdClass;
 use Gettext\Translations;
 use Concrete\Core\Tree\Node\Exception\MoveException;
 
-abstract class Node extends Object implements \Concrete\Core\Permission\ObjectInterface
+abstract class Node extends Object implements \Concrete\Core\Permission\ObjectInterface, AssignableObjectInterface
 {
 
     use AssignableObjectTrait;
@@ -317,11 +318,17 @@ abstract class Node extends Object implements \Concrete\Core\Permission\ObjectIn
         }
     }
 
-    public function executeBeforePermissionAssignment()
+    public function setChildPermissionsToOverride()
     {
-        if (!$this->overrideParentTreeNodePermissions()) {
-            $this->setTreeNodePermissionsToOverride();
+        $this->populateDirectChildrenOnly();
+        foreach($this->getChildNodes() as $child) {
+            $child->setTreeNodePermissionsToOverride();
         }
+    }
+
+    public function setPermissionsToOverride()
+    {
+        $this->setTreeNodePermissionsToOverride();
     }
 
     public function setTreeNodePermissionsToGlobal()
@@ -347,30 +354,32 @@ abstract class Node extends Object implements \Concrete\Core\Permission\ObjectIn
 
     public function setTreeNodePermissionsToOverride()
     {
-        $db = Database::connection();
-        // grab all children
-        $this->populateChildren();
-        $childNodeIDs = $this->getAllChildNodeIDs();
+        if (!$this->overrideParentTreeNodePermissions()) {
+            $db = Database::connection();
+            // grab all children
+            $this->populateChildren();
+            $childNodeIDs = $this->getAllChildNodeIDs();
 
-        $db->executeQuery('delete from TreeNodePermissionAssignments where treeNodeID = ?', [$this->treeNodeID]);
-        // copy permissions from the page to the area
-        $permissions = PermissionKey::getList($this->getPermissionObjectKeyCategoryHandle());
-        foreach ($permissions as $pk) {
-            $pk->setPermissionObject($this);
-            $pk->copyFromParentNodeToCurrentNode();
-        }
+            $db->executeQuery('delete from TreeNodePermissionAssignments where treeNodeID = ?', [$this->treeNodeID]);
+            // copy permissions from the page to the area
+            $permissions = PermissionKey::getList($this->getPermissionObjectKeyCategoryHandle());
+            foreach ($permissions as $pk) {
+                $pk->setPermissionObject($this);
+                $pk->copyFromParentNodeToCurrentNode();
+            }
 
-        $db->executeQuery('update TreeNodes set treeNodeOverridePermissions = 1, inheritPermissionsFromTreeNodeID = ? where treeNodeID = ?', [$this->treeNodeID, $this->treeNodeID]);
-        $this->treeNodeOverridePermissions = true;
-        $this->inheritPermissionsFromTreeNodeID = $this->treeNodeID;
+            $db->executeQuery('update TreeNodes set treeNodeOverridePermissions = 1, inheritPermissionsFromTreeNodeID = ? where treeNodeID = ?', [$this->treeNodeID, $this->treeNodeID]);
+            $this->treeNodeOverridePermissions = true;
+            $this->inheritPermissionsFromTreeNodeID = $this->treeNodeID;
 
-        if (count($childNodeIDs) > 0) {
-            $db->executeQuery(
-                'update TreeNodes set inheritPermissionsFromTreeNodeID = ? where treeNodeID in (' . implode(',', $childNodeIDs) . ') and treeNodeOverridePermissions = 0',
-                [
-                    $this->treeNodeID,
-                ]
-            );
+            if (count($childNodeIDs) > 0) {
+                $db->executeQuery(
+                    'update TreeNodes set inheritPermissionsFromTreeNodeID = ? where treeNodeID in (' . implode(',', $childNodeIDs) . ') and treeNodeOverridePermissions = 0',
+                    [
+                        $this->treeNodeID,
+                    ]
+                );
+            }
         }
     }
